履歴機能つけた

+++ NicoLiveAlertTwitterWPF/ListViewClass/ProgramHistoryListViewData.cs
@@ -9,7 +9,7 @@ namespace NicoLiveAlertTwitterWPF.ListViewClass
     class ProgramHistoryListViewData
     {
         public string Content { get; set; }
-        public int BeginAt { get; set; }
+        public long BeginAt { get; set; }
         public string LiveId { get; set; }
         public string dateTime { get; set; }
     }
